Refactor functions into StringUtil, LogCb
