+++ util/src/test/java/org/datacommons/util/McfUtilTest.java
@@ -15,15 +15,14 @@
 package org.datacommons.util;
 
 import static com.google.common.truth.extensions.proto.ProtoTruth.assertThat;
-import static org.datacommons.util.McfUtil.*;
-import static org.junit.Assert.*;
+import static org.datacommons.util.McfUtil.mergeGraphs;
+import static org.junit.Assert.assertEquals;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
 import java.util.Arrays;
 import java.util.List;
 import org.apache.commons.io.IOUtils;
-import org.datacommons.proto.Debug;
 import org.datacommons.proto.Mcf;
 import org.junit.Rule;
 import org.junit.Test;
@@ -134,81 +133,5 @@ public class McfUtilTest {
                 this.getClass().getResourceAsStream("McfUtilTest_MergedLocations.textproto"),
                 StandardCharsets.UTF_8));
     assertThat(TestUtil.getLocations(act)).ignoringRepeatedFieldOrder().isEqualTo(expLoc);
-  }
-
-  @Test
-  public void funcISO8601Date() {
-    // Year.
-    assertTrue(isValidISO8601Date("2017"));
-    assertFalse(isValidISO8601Date("201"));
-
-    // Year + Month.
-    assertTrue(isValidISO8601Date("2017-01"));
-    assertTrue(isValidISO8601Date("2017-1"));
-    assertTrue(isValidISO8601Date("201701"));
-    assertTrue(isValidISO8601Date("20171"));
-    assertFalse(isValidISO8601Date("2017-Jan"));
-
-    // Year + Month + Day.
-    assertTrue(isValidISO8601Date("2017-1-1"));
-    assertTrue(isValidISO8601Date("2017-11-09"));
-    assertTrue(isValidISO8601Date("20171109"));
-    assertTrue(isValidISO8601Date("2017119"));
-    assertFalse(isValidISO8601Date("2017-Nov-09"));
-
-    // Year + Month + Day + Time.
-    assertTrue(isValidISO8601Date("2017-11-09T22:00"));
-    assertFalse(isValidISO8601Date("2017-11-09D22:00"));
-
-    // Year + Month + Day + Time.
-    assertTrue(isValidISO8601Date("2017-11-09T22:00:01"));
-  }
-
-  @Test
-  public void funcIsNumber() {
-    assertTrue(isNumber("1e10"));
-    assertTrue(isNumber("1.95996"));
-    assertTrue(isNumber("10"));
-    assertTrue(isNumber("-10"));
-    assertTrue(isNumber("-.0010"));
-    assertFalse(isNumber("-.0010x"));
-    assertFalse(isNumber("0xdeadbeef"));
-    assertFalse(isNumber("dc/234"));
-  }
-
-  @Test
-  public void funcIsBool() {
-    assertTrue(isBool("true"));
-    assertTrue(isBool("FALSE"));
-    assertTrue(isBool("1"));
-    assertTrue(isBool("0"));
-    assertFalse(isBool("110"));
-    assertFalse(isBool("yes"));
-    assertFalse(isBool("10"));
-  }
-
-  @Test
-  public void logErrorWithErrCb() {
-    Debug.Log.Builder logCtx = Debug.Log.newBuilder();
-    LogWrapper lw = new LogWrapper(logCtx, testFolder.getRoot().toPath());
-    LogCb logCb = new LogCb(lw, Debug.Log.Level.LEVEL_ERROR, 0);
-    String testCounter = "test_counter";
-    String testMessage = "test_message";
-
-    logCb.logError(testCounter, testMessage);
-    assertTrue(TestUtil.checkLog(logCtx.build(), "test_counter", "test_message"));
-
-    logCb.setDetail(LogCb.VALUE_KEY, "test_value");
-    logCb.logError(testCounter, testMessage);
-    assertTrue(TestUtil.checkLog(logCtx.build(), "test_counter", "value: 'test_value'"));
-
-    logCb.setCounterPrefix("MCF");
-    logCb.logError(testCounter, testMessage);
-    assertTrue(TestUtil.checkLog(logCtx.build(), "MCF_test_counter", "test_message"));
-
-    logCb.setCounterPrefix("");
-    logCb.setCounterSuffix("Prop");
-    logCb.logError(testCounter, testMessage);
-    assertTrue(TestUtil.checkLog(logCtx.build(), "test_counter_Prop", "test_message"));
   }
 }
